Performance improvements to STFReader:  - Simulator..ctor is up to 45% faster.  - RenderProcess.Initialize is up to 73% faster.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@263 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/MSTS/ACTFile.cs
@@ -547,8 +547,8 @@ namespace MSTS
 		public MaxVelocity( STFReader f )
 		{
 			f.VerifyStartOfBlock();
-			A = (float)f.ReadDouble();
-			B = (float)f.ReadDouble();
+			A = f.ReadFloat();
+			B = f.ReadFloat();
 			f.VerifyEndOfBlock();
 		}
 	}
